// src/interfaces/options.interfaces.ts
export interface options {
    whentouse: boolean;
    anatomy: boolean;
    placement: boolean;
    content: boolean;
    behavior: boolean;
    states: boolean;
    interactions: boolean;
}

export interface additionalOptions {
    links: number|string;
    interactivestates: boolean;
    fontconfig: boolean;
    sizevariation: boolean;
    bestpractices: boolean;
    codeexamples: boolean;
}